only hide the #delete_all button when on the comments page

+++ batch-comment-spam-deletion.php
@@ -104,6 +104,10 @@ class PW_BCPD {
 			return;
 		}
 
+		if( 'spam' != $_GET['comment_status'] ) {
+			return;
+		}
+
 		echo '<style>#delete_all { display:none; }</style>';
 
 	}
